activate: use uint64_t for crypt table offset to prevent overflow
get_crypt_table_offset() cast strtoul() to uint32_t before
multiplying by 512, which overflows for sector offsets above
~8M sectors (4GB). Change offset_bytes to uint64_t throughout
the call chain and fix the format strings accordingly.

+++ lib/device/filesystem.h
@@ -33,7 +33,7 @@ struct fs_info {
 	char fs_dev_path[PATH_MAX]; /* usually lv dev, can be crypt dev */
 	unsigned int fs_block_size_bytes; /* 512 or 4k */
 	uint64_t fs_last_byte; /* last byte on the device used by the fs */
-	uint32_t crypt_offset_bytes; /* offset in bytes of crypt data on LV */
+	uint64_t crypt_offset_bytes; /* offset in bytes of crypt data on LV */
 	dev_t crypt_devt; /* dm-crypt device between the LV and FS */
 	uint64_t crypt_dev_size_bytes;
